+++ packages/terra-text/src/terra-dev-site/test/text/TextWordWrapped.test.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import Text from '../../../Text';
+
+export default () => <Text id="text" isWordWrapped>LongTextShouldBreakWordAndWrapLalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalalala</Text>;