+++ packages/terra-text/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added isWordWrapped prop
 
+### Removed
+* Removed dependency on terra-heading
+
 2.29.1 - (November 20, 2018)
 ------------------
 ### Changed