+++ packages/terra-text/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Removed
+* Removed dependency on terra-heading
 
 2.29.1 - (November 20, 2018)
 ------------------