+++ packages/terra-text/CHANGELOG.md
@@ -3,8 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+2.30.0 - (January 2, 2019)
+------------------
 ### Added
 * Added isWordWrapped prop
+
+### Changed
+* Minor dependency version bump
 
 ### Removed
 * Removed dependency on terra-heading